+++ srcs/parser/parser_all.c
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/13 14:24:09 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/04/01 10:46:28 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/04/01 14:30:50 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ t_error *parser_all(t_env *e)
         return (err);
     if ((err = parser_room(e, &line)) != NULL)
         return (err);
-    if (e->start == (unsigned int)-1 || e->end == (unsigned int)-1)
+    if (e->start == (size_t)-1 || e->end == (size_t)-1)
         return (error_create("Start and End command are needed.", NULL, 15));
     if ((e->graph = graph_new(e->map->length)) == NULL)
         return (error_create("Graph malloc failed.", NULL, 9));
