+++ srcs/parser/parser_room.c
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/21 20:17:08 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/03/31 23:32:10 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/04/01 14:31:24 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,9 +42,9 @@ t_error *parser_room(t_env *e, char **line)
         if (*line[0] == '#')
         {
             type_change(*line, &type);
-            if (type == START && e->start != (unsigned int)-1)
+            if (type == START && e->start != (size_t)-1)
                 err = error_create("Start is in double.", NULL, 12);
-            if (type == END && e->end != (unsigned int)-1)
+            if (type == END && e->end != (size_t)-1)
                 err = error_create("End is in double.", NULL, 13);
         }
         else if (*line[0] == 'L')
